mailer refactorings, fixed messenger bug
